Arithmetic and Statistical Operations on Persistence Landscapes

Calculate sums, scalar multiples, absolute values, means, inner products, extrema, and moments of persistent landscapes. These operations arise from the Hilbert space structure on persistence landscapes (Bubenik, 2015).

pl_add(pl1, pl2)

pl_sum(pl_list)

pl_scale(pl, mult)

pl_abs(pl)

pl_diff(pl_list)

pl_mean(pl_list)

pl_var(pl_list, p = 2)

pl_sd(pl_list, p = 2)

pl_inner(pl1, pl2)

pl_min(pl, level = 1L)

pl_max(pl, level = 1L)

pl_range(pl, level = 1L)

pl_vmin(pl, level = pl_num_levels(pl))

pl_vmax(pl, level = pl_num_levels(pl))

pl_vrange(pl, level = pl_num_levels(pl))

pl_moment(pl, p = 1L, center = 0, level = 1L)

pl_vmoment(pl, p = 1L, center = 0, level = pl_num_levels(pl))

Arguments

pl1, pl2

Persistent landscapes.

pl_list

A list of persistent landscapes.

pl

Persistent landscapes.

mult

Double; a real-valued scale factor.

p

Positive integer or infinity; the power used to compute a norm or moment.

level

Positive integer; the level of the persistence landscape (up to) whose moment to calculate.

center

Double; where to center the moment.

Value

A persistence landscape (an object of S4 class 'Rcpp_PersistenceLandscape'), a real number, or a vector of real numbers.

Details

These functions are prefixed pl_*() to help users access them via tab-completion. Some take their names from the underlying S4 class methods and are only provided to enable composition via pipes: add, scale, abs, inner, min (minimum), max (maximum), and moment; range combines min and max. Others mimic classic R functions to handle lists of persistence landscapes: sum, diff, mean, var, and sd. Finally, some are vectorizations of the preceding: vmin, vmax, vrange, and vmoment.

See also

Examples

# scale a landscape
x <- tdaunif::sample_torus_tube(4, 2.5)
pd <- as_persistence(ripserr::vietoris_rips(x, dim = 1L, threshold = 2))
#> Warning: `dim` parameter has been deprecated; use `max_dim` instead.
pl <- landscape(pd$pairs[[1]], exact = FALSE, xmax = 2.5, by=0.1)
print(pl$getInternal()[2, , ])
#>       [,1]       [,2]
#>  [1,]  0.0 0.00000000
#>  [2,]  0.1 0.10000000
#>  [3,]  0.2 0.20000000
#>  [4,]  0.3 0.30000000
#>  [5,]  0.4 0.40000000
#>  [6,]  0.5 0.41779648
#>  [7,]  0.6 0.31779648
#>  [8,]  0.7 0.21779648
#>  [9,]  0.8 0.11779648
#> [10,]  0.9 0.01779648
#> [11,]  1.0 0.00000000
#> [12,]  1.1 0.00000000
#> [13,]  1.2 0.00000000
#> [14,]  1.3 0.00000000
#> [15,]  1.4 0.00000000
#> [16,]  1.5 0.00000000
#> [17,]  1.6 0.00000000
#> [18,]  1.7 0.00000000
#> [19,]  1.8 0.00000000
#> [20,]  1.9 0.00000000
#> [21,]  2.0 0.00000000
#> [22,]  2.1 0.00000000
#> [23,]  2.2 0.00000000
#> [24,]  2.3 0.00000000
#> [25,]  2.4 0.00000000
#> [26,]  2.5 0.00000000
print(pl$scale(0.5)$getInternal()[2, , ])
#>       [,1]        [,2]
#>  [1,]  0.0 0.000000000
#>  [2,]  0.1 0.050000000
#>  [3,]  0.2 0.100000000
#>  [4,]  0.3 0.150000000
#>  [5,]  0.4 0.200000000
#>  [6,]  0.5 0.208898239
#>  [7,]  0.6 0.158898239
#>  [8,]  0.7 0.108898239
#>  [9,]  0.8 0.058898239
#> [10,]  0.9 0.008898239
#> [11,]  1.0 0.000000000
#> [12,]  1.1 0.000000000
#> [13,]  1.2 0.000000000
#> [14,]  1.3 0.000000000
#> [15,]  1.4 0.000000000
#> [16,]  1.5 0.000000000
#> [17,]  1.6 0.000000000
#> [18,]  1.7 0.000000000
#> [19,]  1.8 0.000000000
#> [20,]  1.9 0.000000000
#> [21,]  2.0 0.000000000
#> [22,]  2.1 0.000000000
#> [23,]  2.2 0.000000000
#> [24,]  2.3 0.000000000
#> [25,]  2.4 0.000000000
#> [26,]  2.5 0.000000000

# create two landscapes from the same sampling distribution
x <- tdaunif::sample_torus_tube(4, 2.5)
y <- tdaunif::sample_torus_tube(4, 2.5)
x_pd <- as_persistence(ripserr::vietoris_rips(x, dim = 1, threshold = 2))
#> Warning: `dim` parameter has been deprecated; use `max_dim` instead.
y_pd <- as_persistence(ripserr::vietoris_rips(y, dim = 1, threshold = 2))
#> Warning: `dim` parameter has been deprecated; use `max_dim` instead.
x_pl <- landscape(x_pd$pairs[[1]], exact = FALSE, xmax = 2.5, by = 0.1)
y_pl <- landscape(y_pd$pairs[[1]], exact = FALSE, xmax = 2.5, by = 0.1)

# compare landscapes and calculate mean landscape
print(x_pl$getInternal())
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0  0.1  0.2  0.3  0.4  0.5  0.6  0.7  0.8   0.9     1   1.1   1.2   1.3
#> [2,]    0  0.1  0.2  0.3  0.4  0.5  0.6  0.7  0.8   0.9     1   1.1   1.2   1.3
#> [3,]    0  0.1  0.2  0.3  0.4  0.5  0.6  0.7  0.8   0.9     1   1.1   1.2   1.3
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]   1.4   1.5   1.6   1.7   1.8   1.9     2   2.1   2.2   2.3   2.4   2.5
#> [2,]   1.4   1.5   1.6   1.7   1.8   1.9     2   2.1   2.2   2.3   2.4   2.5
#> [3,]   1.4   1.5   1.6   1.7   1.8   1.9     2   2.1   2.2   2.3   2.4   2.5
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]      [,5]      [,6]      [,7]       [,8]      [,9]
#> [1,]    0  0.1  0.2  0.3 0.4000000 0.5000000 0.6000000 0.70000000 0.8000000
#> [2,]    0  0.1  0.2  0.3 0.4000000 0.5000000 0.5385641 0.43856414 0.3385641
#> [3,]    0  0.1  0.2  0.3 0.3548517 0.2548517 0.1548517 0.05485168 0.0000000
#>          [,10]     [,11]      [,12]    [,13]    [,14]    [,15]    [,16]
#> [1,] 0.8143040 0.7143040 0.61430402 0.514304 0.414304 0.314304 0.214304
#> [2,] 0.2385641 0.1385641 0.03856414 0.000000 0.000000 0.000000 0.000000
#> [3,] 0.0000000 0.0000000 0.00000000 0.000000 0.000000 0.000000 0.000000
#>         [,17]      [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,] 0.114304 0.01430402     0     0     0     0     0     0     0     0
#> [2,] 0.000000 0.00000000     0     0     0     0     0     0     0     0
#> [3,] 0.000000 0.00000000     0     0     0     0     0     0     0     0
#> 
print(y_pl$getInternal())
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0  0.1  0.2  0.3  0.4  0.5  0.6  0.7  0.8   0.9     1   1.1   1.2   1.3
#> [2,]    0  0.1  0.2  0.3  0.4  0.5  0.6  0.7  0.8   0.9     1   1.1   1.2   1.3
#> [3,]    0  0.1  0.2  0.3  0.4  0.5  0.6  0.7  0.8   0.9     1   1.1   1.2   1.3
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]   1.4   1.5   1.6   1.7   1.8   1.9     2   2.1   2.2   2.3   2.4   2.5
#> [2,]   1.4   1.5   1.6   1.7   1.8   1.9     2   2.1   2.2   2.3   2.4   2.5
#> [3,]   1.4   1.5   1.6   1.7   1.8   1.9     2   2.1   2.2   2.3   2.4   2.5
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3]      [,4]       [,5]      [,6]      [,7]      [,8]
#> [1,]    0  0.1  0.2 0.3000000 0.40000000 0.5000000 0.6000000 0.5347691
#> [2,]    0  0.1  0.2 0.3000000 0.40000000 0.3288382 0.2288382 0.1288382
#> [3,]    0  0.1  0.2 0.1589957 0.05899566 0.0000000 0.0000000 0.0000000
#>            [,9]     [,10]     [,11]     [,12]      [,13] [,14] [,15] [,16]
#> [1,] 0.43476914 0.3347691 0.2347691 0.1347691 0.03476914     0     0     0
#> [2,] 0.02883823 0.0000000 0.0000000 0.0000000 0.00000000     0     0     0
#> [3,] 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000     0     0     0
#>      [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]     0     0     0     0     0     0     0     0     0     0
#> [2,]     0     0     0     0     0     0     0     0     0     0
#> [3,]     0     0     0     0     0     0     0     0     0     0
#> 
print(pl_mean(list(x_pl, y_pl))$getInternal())
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0  0.1  0.2  0.3  0.4  0.5  0.6  0.7  0.8   0.9     1   1.1   1.2   1.3
#> [2,]    0  0.1  0.2  0.3  0.4  0.5  0.6  0.7  0.8   0.9     1   1.1   1.2   1.3
#> [3,]    0  0.1  0.2  0.3  0.4  0.5  0.6  0.7  0.8   0.9     1   1.1   1.2   1.3
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]   1.4   1.5   1.6   1.7   1.8   1.9     2   2.1   2.2   2.3   2.4   2.5
#> [2,]   1.4   1.5   1.6   1.7   1.8   1.9     2   2.1   2.2   2.3   2.4   2.5
#> [3,]   1.4   1.5   1.6   1.7   1.8   1.9     2   2.1   2.2   2.3   2.4   2.5
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3]      [,4]      [,5]      [,6]       [,7]       [,8]
#> [1,]    0  0.1  0.2 0.3000000 0.4000000 0.5000000 0.60000000 0.61738457
#> [2,]    0  0.1  0.2 0.3000000 0.4000000 0.4144191 0.38370118 0.28370118
#> [3,]    0  0.1  0.2 0.2294978 0.2069237 0.1274258 0.07742584 0.02742584
#>           [,9]     [,10]      [,11]      [,12]     [,13]    [,14]    [,15]
#> [1,] 0.6173846 0.5745366 0.47453658 0.37453658 0.2745366 0.207152 0.157152
#> [2,] 0.1837012 0.1192821 0.06928207 0.01928207 0.0000000 0.000000 0.000000
#> [3,] 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000 0.000000 0.000000
#>         [,16]      [,17]       [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#> [1,] 0.107152 0.05715201 0.007152012     0     0     0     0     0     0     0
#> [2,] 0.000000 0.00000000 0.000000000     0     0     0     0     0     0     0
#> [3,] 0.000000 0.00000000 0.000000000     0     0     0     0     0     0     0
#>      [,26]
#> [1,]     0
#> [2,]     0
#> [3,]     0
#> 

if (FALSE) {
set.seed(492869L)

# compute landscape for a large sample
pt <- tdaunif::sample_torus_tube(1000, 5)
pd <- as_persistence(ripserr::vietoris_rips(pt, dim = 2, threshold = 2))
pl <- landscape(pd, degree = 1, exact = FALSE, by = 0.1, xmin = 0, xmax = 2)

# compute landscapes for a large sample of small samples
pl_list <- c()
for (i in seq(100)) {
  pti <- tdaunif::sample_torus_tube(100, 5)
  pdi <- as_persistence(ripserr::vietoris_rips(pti, dim = 2, threshold = 2))
  pli <- landscape(pdi, degree = 1, exact = FALSE, by = 0.1, xmin = 0, xmax = 2)
  pl_list <- c(pl_list, pli)
}

# compute the mean landscape
pl_avg <- pl_mean(pl_list)

# compute the distance between the landscapes
pl_diff <- pl$add(pl_avg$scale(-1))
print(pl_inner(pl_diff, pl_diff))
}